requirements.txt
langchain>=1.2.15,<2
langchain-classic>=1.0.3,<2
langchain-community>=0.4,<1
langchain-ollama>=1.1,<2
langchain-google-genai>=4.2,<5
langchain-anthropic>=1.4,<2
langchain-huggingface>=1.2,<2
langchain-redis>=0.2,<1
redis>=5,<7.2
sentence-transformers>=3,<6
unstructured[pdf]>=0.16,<0.17
python-dotenv>=1,<2
chardet>=3,<6
pyyaml>=6,<7
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1